Move filters to main elasticsearch class
